Use try-with-resources on ZIP File
The problem why the ZIP file would be closed during the read of it's
entries is, that we previously returned a stream of SQ's rather than a
completely build list. The stream is not realized as we had returned it,
so the ZIP file will be closed before the individual stream items would
be processed. Now fully build the list inside the try-with-resources
block.